fix db integer sizes

+++ app/Stubs/Migrations/HasFulfilmentDelivery.php
@@ -18,7 +18,7 @@ trait HasFulfilmentDelivery
         $table = $this->groupOrgRelationship($table);
         $table->string('slug')->unique()->collation('und_ns');
         $table->ulid()->unique()->index();
-        $table->unsignedSmallInteger('fulfilment_customer_id');
+        $table->unsignedInteger('fulfilment_customer_id');
         $table->foreign('fulfilment_customer_id')->references('id')->on('fulfilment_customers');
         $table->unsignedSmallInteger('fulfilment_id');
         $table->foreign('fulfilment_id')->references('id')->on('fulfilments');
